refactor: extract shared Core Audio helpers into audio_util
